Move GC unknown outbound routes telemetry to telemetry tracker
